pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.14",
    "langchain-huggingface>=0.1.2",
    "openai>=1.59.7",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.3.1",
]

[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "chromadb>=0.6.3",
    "jupyterlab>=4.3.4",
    "langchain-community>=0.3.14",
    "langchain-text-splitters>=0.3.5",
    "pypdf>=5.1.0",
]
